Modifier bleeding problem solved, with caveats.
Closes #4.

This only really works under the assumption that 
there are only two keys to deal with, and that 
they are the result of an overlap caused by fast  typing (as one key is being released, another is  being pressed). There is no way (short of writing  a custom keyboard driver) of correctly handling  multiple simultaneous keys when some have  modifiers and some don't.  This is a limitation of  the HID keyboard protocol.